--- tasklist.txt
@@ -1,4 +1,0 @@
-3
-Game|Play game|2|0|2025-6-24
-Coding|do code|2|0|2025-6-30
-Clean|-|0|0|2025-6-25
