--- test.txt
@@ -1,2 +1,0 @@
-This is the main branch. I have now entered the test branch.
-
